feat: adds division into modules
